lints and pulls all styled components
